Added unsubscribe functionality
Since the Mongo Object Ids are really hard to be guessed, I think it is fine to use them in the unsubscribe process.
I added a route: /api/unsubscribe?id=<objectid>. This is added at the end of the emails.

This endpoint just sets the emailSubscription to false and then redirect to the home page.